Make getFloatTimeSinceEpoch more precise to bring back smoothness during cutscenes
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@11647 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/modes/world_status.hpp
@@ -91,7 +91,7 @@ protected:
     /**
       * Elasped/remaining time in seconds
       */
-    float           m_time;
+    double          m_time;
     ClockType       m_clock_mode;
 
     bool            m_play_racestart_sounds;
